Add grouping "vaccine for platform"
Similarly to #620, this adds a mid-level grouping term between VO:0000001 (vaccine) and several children representing vaccine platforms:

1. conjugate vaccine
2. DNA vaccine
3. inactivated vaccine
4. live virulent vaccine
5. live attenuated vaccine
6. passive vaccine
7. toxoid vaccine
8. whole organism vaccine
9. virus-like particle vaccine
10. recombinant vector vaccine
11. RNA vaccine
12. subunit vaccine

This is added in the term "vaccine for platform" (VO:0005505). This doesn't add any axiomization regarding this grouping at the moment, since it's not yet clear if there's a shared way of ontologizing "what is a platform". Looking forward, platforms are typically defined as a combination of the one or more of the components/antigens/adjuvants/other ingredient types